+++ app/helpers/rad_common/application_helper.rb
@@ -117,22 +117,6 @@ module RadCommon
       Regexp.new regex_str
     end
 
-    def enum_to_translated_option(record, enum_name)
-      RadEnum.new(record.class, enum_name).translated_option(record)
-    end
-
-    def options_for_enum(klass, enum_name)
-      RadEnum.new(klass, enum_name).options
-    end
-
-    def options_for_enum_reversed(klass, enum_name)
-      RadEnum.new(klass, enum_name).options_reversed
-    end
-
-    def enum_translation(klass, enum_name, value)
-      RadEnum.new(klass, enum_name).translation(value)
-    end
-
     def bootstrap_flash
       flash_messages = []
