+++ spec/helpers/rad_common/application_helper_spec.rb
@@ -3,7 +3,6 @@ require 'rails_helper'
 describe RadCommon::ApplicationHelper do
   let(:me) { create :user }
   let(:date) { Time.current }
-  let(:division) { create :division }
   let(:timestamp) { '2018-06-15 06:43 AM' }
 
   describe '#show_actions?' do
@@ -45,41 +44,6 @@ describe RadCommon::ApplicationHelper do
       it 'returns false' do
         expect(helper.show_actions?(model_class)).to be(false)
       end
-    end
-  end
-
-  describe 'enum_to_translated_option' do
-    let(:error_message) { "enum division_status_xxx on Division doesn't exist" }
-
-    it 'translates the value' do
-      expect(enum_to_translated_option(division, :division_status)).to eq 'Active'
-    end
-
-    it 'handles nil' do
-      division.division_status = nil
-      expect(enum_to_translated_option(division, :division_status)).to be_nil
-    end
-
-    it 'handles blank' do
-      division.division_status = ''
-      expect(enum_to_translated_option(division, :division_status)).to be_nil
-    end
-
-    it 'raises error when missing enum' do
-      expect { enum_to_translated_option(division, :division_status_xxx) }.to raise_error(error_message)
-    end
-  end
-
-  describe 'options_for_enum' do
-    subject { options_for_enum(Division, :division_status) }
-
-    let(:options) { [%w[Pending status_pending], %w[Active status_active], %w[Inactive status_inactive]] }
-    let(:error_message) { "enum division_status_xxx on Division doesn't exist" }
-
-    it { is_expected.to eq options }
-
-    it 'raises error when missing enum' do
-      expect { options_for_enum(Division, :division_status_xxx) }.to raise_error(error_message)
     end
   end
